requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
django-password-validators
Pillow
django-filter